requirements.txt
Django==1.10.6
python-dateutil
